formatting file's functions
